Move components in prep for merge
